+++ CHANGELOG.md
@@ -44,6 +44,8 @@
   [#462](https://github.com/Kong/gateway-operator/pull/462)
 - Add `KongService` reconciler for Konnect control planes.
   [#470](https://github.com/Kong/gateway-operator/pull/470)
+- Add `KongUpstream` reconciler for Konnect control planes.
+  [#593](https://github.com/Kong/gateway-operator/pull/593)
 - Add `KongConsumer` reconciler for Konnect control planes.
   [#493](https://github.com/Kong/gateway-operator/pull/493)
 - Add `KongRoute` reconciler for Konnect control planes.